backend/requirements.txt
# FastAPI核心依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Google Ads API
google-ads==28.4.0

# 定时任务
apscheduler==3.10.4

# 工具库
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0

# 日期时间处理
python-dateutil==2.8.2

# CORS支持
fastapi-cors==0.0.6
